+++ src/emu/nosv/breakdown.c
@@ -23,6 +23,7 @@
 #include "recorder.h"
 #include "sort.h"
 #include "system.h"
+#include "thread.h"
 #include "task.h"
 #include "track.h"
 #include "value.h"
@@ -42,6 +43,28 @@ create_cpu(struct bay *bay, struct nosv_breakdown_cpu *bcpu, int64_t gindex)
 	}
 	if (bay_register(bay, &bcpu->tri) != 0) {
 		err("bay_register tri failed");
+		return -1;
+	}
+
+	return 0;
+}
+
+static int
+check_thread_metadata(struct thread *th)
+{
+	if (th->meta == NULL) {
+		err("thread has no metadata");
+		return -1;
+	}
+
+	JSON_Value *val = json_object_dotget_value(th->meta, "nosv.can_breakdown");
+	if (val == NULL) {
+		err("missing nosv.can_breakdown attribute");
+		return -1;
+	}
+
+	if (!json_value_get_boolean(val)) {
+		err("nosv.can_breakdown is false, missing events to enable breakdown");
 		return -1;
 	}
 
@@ -84,6 +107,13 @@ model_nosv_breakdown_create(struct emu *emu)
 
 		if (create_cpu(&emu->bay, bcpu, cpu->gindex) != 0) {
 			err("create_cpu failed");
+			return -1;
+		}
+	}
+
+	for (struct thread *th = emu->system.threads; th; th = th->gnext) {
+		if (check_thread_metadata(th) != 0) {
+			err("bad nosv metadata in thread: %s", th->id);
 			return -1;
 		}
 	}