+++ src/emu/nosv/setup.c
@@ -82,7 +82,7 @@ static struct ev_decl model_evlist[] = {
 
 struct model_spec model_nosv = {
 	.name    = model_name,
-	.version = "2.2.0",
+	.version = "2.3.0",
 	.evlist  = model_evlist,
 	.model   = model_id,
 	.create  = model_nosv_create,